Add automatic pushing to indexing queue as part of high score import process
