+++ osu.Server.Queues.ScoreStatisticsProcessor/ScoreStatisticsProcessor.cs
@@ -161,28 +161,5 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor
 
             return rulesetsToProcess;
         }
-
-        private class ElasticQueueProcessor : QueueProcessor<ElasticQueueProcessor.ElasticScoreItem>
-        {
-            private static readonly string queue_name = $"score-index-{Environment.GetEnvironmentVariable("SCHEMA")}";
-
-            internal ElasticQueueProcessor()
-                : base(new QueueConfiguration { InputQueueName = queue_name })
-            {
-                // TODO: automate schema version lookup
-                // see https://github.com/ppy/osu-elastic-indexer/blob/316e3e2134933e22363f4911e0be4175984ae15e/osu.ElasticIndexer/Redis.cs#L10
-            }
-
-            protected override void ProcessResult(ElasticScoreItem scoreItem)
-            {
-                throw new NotImplementedException();
-            }
-
-            [Serializable]
-            public class ElasticScoreItem : QueueItem
-            {
-                public long? ScoreId { get; init; }
-            }
-        }
     }
 }
